% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_backup.R
\name{d_backup}
\alias{d_backup}
\title{Backup a data base}
\usage{
d_backup(fsq)
}
\arguments{
\item{fsq}{Full name of the data base}
}
\value{
A saved data base with extension .bak
}
\description{
Back a htsr sqlite data base
}
\author{
P. Chevallier - Jan 2019 / Nov 2020
}
