% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_compact.R
\name{d_compact}
\alias{d_compact}
\title{Compact a data base}
\usage{
d_compact(fsq)
}
\arguments{
\item{fsq}{Full name of the data base}
}
\value{
New data base or overwritten data base. Note that the created data base is empty.
}
\description{
Compact htsr sqlite data base
}
\author{
P. Chevallier - Jan 2019
}
