% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_convert_eaufrance.R
\name{d_convert_eaufrance}
\alias{d_convert_eaufrance}
\title{Convert eaufrance station files into a htsr sqlite base}
\usage{
d_convert_eaufrance(eaufrance.dir, station.id, fsqname)
}
\arguments{
\item{eaufrance.dir}{Full path of the folder were the eaufrance data base is extracted (character)}

\item{station.id}{Id list of the stations to convert (character)}

\item{fsqname}{Name of the returned sqlite data base without extension (character)}
}
\value{
A sqlite database compatible with the htsr library.
}
\description{
Convert a eaufrance  hydrological file into a htsr sqlite base. It regards
the "basic" data file, which includes water level and discharge data. .
}
\details{
The data base is build from selected stations in the "stations.tar" file available
on the data.eaufrance web site : \url{https://data.ofb.fr/catalogue/data-eaufrance/fre/catalog.search}.
This file must be first downloaded and extracted in the folder eaufrance.dir. For the extraction the R
function untar() can be used.

Secondly, within the eaufrance.dir, the file stations/stations.csv give the full list of the available stations.
One or more station ids
must be chosen and included in the station.id list parameter.
Another possibility is to consult the "Hydro Portail"
(\url{https://www.hydro.eaufrance.fr/rechercher/entites-hydrometriques}) in order
to select the station ids.

In the sqlite data base, the units of water level data is cm and of discharge data is m3/s.
}
\author{
P. Chevallier - Jul/Aug 2024
}
