% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_convert_weewx.R
\name{d_convert_weewx}
\alias{d_convert_weewx}
\title{Convert a weewx data base into a htsr sqlite base}
\usage{
d_convert_weewx(db.sqlite, db.weewx, sta, name_st, tzo = "CET", bku = TRUE)
}
\arguments{
\item{db.sqlite}{Full name of the htsr data base}

\item{db.weewx}{Full name of the weewx data base}

\item{sta}{Station id}

\item{name_st}{Station name}

\item{tzo}{Time zone of the weewx data (default = "CET")}

\item{bku}{Backup the data base (default = TRUE)}
}
\description{
Convert a weewx data base into a tshm sqlite base
}
\examples{
\dontrun{

d_imp_weewx("foo.sqlite","weewx.sql")
}

}
\seealso{
\code{\link{d_inventory}} or \code{\link{ds_inventory}} list the content
of the data base ;
 \code{\link{d_exp_hts}} to extract time-series.
}
\author{
P. Chevallier - Feb 2018 - dec 2019
}
