% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_create.R
\name{d_create}
\alias{d_create}
\title{Create a data base}
\usage{
d_create(db.sqlite, cr_table = TRUE, bku = TRUE)
}
\arguments{
\item{db.sqlite}{Full name of the data base}

\item{cr_table}{Create the basis tables : TRUE (default), FALSE}

\item{bku}{Operate a backup if db.sqlite exists : TRUE (default) / FALSE}
}
\value{
a new data base
}
\description{
Create htsr sqlite data base
}
\details{
If the data base already exists and bku is TRUE, a backup is
automatically generated.

If cr_table is TRUE, The following tables are also created:
ST (stations), SS (sensors), WL (water levels), DI (discharges), PR
(Precipitations), WE (weather) and QU (quality)
}
\author{
P. Chevallier - Jan 2019
}
