% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_exphts.R
\name{d_exphts}
\alias{d_exphts}
\title{Shiny app: export hts files from a sqlite data base}
\usage{
d_exphts(fsq)
}
\arguments{
\item{fsq}{File name of the Sqlite data base}
}
\value{
a shiny session
}
\description{
Shiny application of the \code{\link{d_exp_hts}} function
}
\details{
When launched, this function first ask to choose the sqlite data base.
 Complete the requested information in the left panel, then press the submit button in
order to extract the file. If you want to display the plot of the extacted file,
choose "line" or "bar" and press the plot button. When finished press "done".

If the data do not exist, the app crashes and error messages
are displayed in the console window.
}
\author{
P. Chevallier - Apr-Nov 2020
}
