% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_rem_hts.R
\name{d_rem_hts}
\alias{d_rem_hts}
\title{Remove hts records from a data base}
\usage{
d_rem_hts(db.sqlite, table, sta, sen, start = NA, end = NA)
}
\arguments{
\item{db.sqlite}{Full name of the data base}

\item{table}{Table}

\item{sta}{Station id}

\item{sen}{Sensor id}

\item{start}{Start time of removed records}

\item{end}{End time of removed records}
}
\value{
Actualized data base
}
\description{
Remove hst records from a Sqlite base
}
\details{
The main table where the data have to be removed must be selected with one the
following abbreviation: WL (water level), DI (discharge), WE (weather), PR (precipitation)
or QU (quality)
}
\author{
P. Chevallier - jan 2019 - nov 2020
}
