% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_table.R
\name{d_table}
\alias{d_table}
\title{Create or remove a table of a htsr sqlite base}
\usage{
d_table(fsq, table, op = "C", bku = TRUE)
}
\arguments{
\item{fsq}{Full name of the data base}

\item{table}{Table name}

\item{op}{Create (default) or Remove C/R}

\item{bku}{Automatic Backup TRUE (default) / FALSE}
}
\value{
Table created or removed
}
\description{
The function allows to create or remove of a tshm sqlite base. If
the base doesn't exist, it is created.
}
\details{
Possible table names : ST (Stations), SS (Sensors), WL (Water levels), DI (Discharges),
WE, (Weather), PR (Precipitations), QU (Quality)
}
\seealso{
\itemize{
 \item \code{\link{ds_inventory}} to list the content of the base ;
 \item \code{\link{ds_exp_hts}} to extract a time-series
}
}
\author{
P. Chevallier - Jan-Feb 2018
}
