% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_wind.R
\name{d_wind}
\alias{d_wind}
\title{Create a wind table}
\usage{
d_wind(db.sqlite, sta = NA, swd = NA, swv = NA)
}
\arguments{
\item{db.sqlite}{Full name of the tshm data base}

\item{sta}{Station id}

\item{swd}{Id of wind direction sensor}

\item{swv}{Id of wind speed sensor}
}
\value{
A tibble named "data_wind" with 5 columns date, month, year, wind_dir, wind_spd
}
\description{
Create a tibble with wind direction and speed
}
\examples{
\dontrun{

h_wind (db.sqlite, sta="VB", swd="WD", swv="WV")
}

}
\seealso{
\code{\link{p_wind}} plot wind roses
}
\author{
P. Chevallier - Dec 2019
}
