% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_dismeas.R
\name{ds_dismeas}
\alias{ds_dismeas}
\title{Add, Modify or Remove discharge measurements (Shiny app)}
\usage{
ds_dismeas(fsq, sta, sen)
}
\arguments{
\item{fsq}{htsr data base}

\item{sta}{Station Id.}

\item{sen}{Sensor Id.}
}
\value{
an actualized data base
}
\description{
Add, Modify or Remove discharge measurements for a station/sensor
}
\author{
P. Chevallier - Dec 2020
}
