% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_exp_hts.R
\name{ds_exp_hts}
\alias{ds_exp_hts}
\title{Extraction of a time-series from htsr data base}
\usage{
ds_exp_hts()
}
\value{
The function returns a file (nomfic) with the following name: <sensor.id>_<station.id>.hts
}
\description{
The function display a web page allowing to extract a time-series in the "hts" format.
}
\details{
Complete the requested information in the left panel, then press the submit button in
order to extract the file. If you want to display the plot of the extracted file,
choose "line" or "bar" and press the plot button.

 When the subfunction "d_exp_hts(fsq, sta,sen,rtime=FALSE,dstart=NA,dend=NA, rplot=FALSE)"
 is used solely it returns a tibble tstab with 4 columns Date, Value, Station, Sensor.
 In this last subfunction fsq is the sqlite data base; sta, the station id, sen, the sensor id; rtime, dstart and
 dend define a time interval; rplot, the resulted plot.
}
\author{
P. Chevallier - Oct 2017 - Sep 2023
}
