% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_inventory_station.R
\name{ds_inventory_station}
\alias{ds_inventory_station}
\title{Inventory of the stations of an htsr data base}
\usage{
ds_inventory_station()
}
\value{
A table with the inventory of stations in the data base.
}
\description{
The function display a web page in order to produce an inventory of the
stations in an htsr data base.
}
\details{
Complete the requested information in the left panel, then press the submit button.
 When finished press "done".

 If "Output format" is "none", the results are displayed on the screen, If it is "xlsx",
 or "csv" (, as separator) or "csv2" (; as separator), the corresponding file
 with a station list is written.
}
\author{
P. Chevallier - Dec 2023
}
