% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_station.R
\name{ds_station}
\alias{ds_station}
\title{Create, Modify or Remove a station}
\usage{
ds_station()
}
\value{
Station created, modified ou removed from the data base
}
\description{
Create, Modify or Remove a station.
}
\details{
If operation is Create, the fields Id_Station, Type_Station and
Name (name_st) are compulsory. The field Name can be modified afterwards.


}
\author{
P. Chevallier - Jan 2018 - Sep 2023
}
