% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_change_id.R
\name{f_change_id}
\alias{f_change_id}
\title{Change Station id or Sensor id in a hts file}
\usage{
f_change_id(file, sta = NA, sen = NA, overwrite = FALSE)
}
\arguments{
\item{file}{file to proceed}

\item{sta}{new station id (default: NA)}

\item{sen}{new sensor id (default: NA)}

\item{overwrite}{TRUE / FALSE (default) if the output file exists}
}
\description{
The function changes the station and/or the sensor id of a hts file.
The new file is renamed with the new ids and a prefix n_: nw_<sensor.id>_<station.id>.hts, BUT the
eventual prefixes or suffixes of the original name are not conserved.
The original file is not removed.
}
\author{
P. Chevallier - Nov 2017-Jan 2019
}
