% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_month2day.R
\name{f_month2day}
\alias{f_month2day}
\title{Interpolation of daily records from a monthly time series}
\usage{
f_month2day(file)
}
\arguments{
\item{file}{monthly time series to process}
}
\value{
a daily time series
}
\description{
Interpolation of daily records from a monthly time-series
}
\details{
The function build and interpolated daily time-series from a monthly one. The 
daily values are linearly computed between two consecutive monthly values.
}
\author{
P. Chevallier - dec 2022
}
