% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_properties.R
\name{f_properties}
\alias{f_properties}
\title{Properties of a hts series}
\usage{
f_properties(file, gaps = FALSE)
}
\arguments{
\item{file}{: file to be analyzed}

\item{gaps}{: produce a file with a table of the gaps: TRUE / FALSE (par defaut)}
}
\value{
Basic infos on a hts time-series
}
\description{
The function provides the properties of a time-series, its duration and
the inventory of its gaps
}
\details{
If gaps = TRUE, a file is produced, with the same name of file and the extension .gap.
It contents a table with the gaps of the series and allows to build a plot with the
function \code{\link{p_gaps}}.
}
\seealso{
\code{\link{p_gaps}}.
}
\author{
P. Chevallier - January 2019
}
