% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_addna.R
\name{h_addna}
\alias{h_addna}
\title{Add NA values within a time series}
\usage{
h_addna(file, add)
}
\arguments{
\item{file}{File name to proceed}

\item{add}{List of dates with NA values to be added}
}
\description{
Add NA values within a time series
}
\details{
The function adds records with NA in a time series at given dates. 
If the date already exists, the value is replaced by NA

The output file is named with a nap_ prefix.
}
\examples{
\dontrun{

f <- h_addna (f, add = c("2021-01-01 12:00:00 UTC", "2031-01-01 12:00:00 UTC"))
}

}
\author{
P. Chevallier - November 2022
}
