% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_adjust.R
\name{h_adjust}
\alias{h_adjust}
\title{Adjust a time series to a statistical model}
\usage{
h_adjust(file, time_unit = "year")
}
\arguments{
\item{file}{File to proceed}

\item{time_unit}{to be chosen in: "100y", "year", "month", "day"}
}
\description{
Adjust a time series to a statistical model
}
\details{
The function adjust a time series with a statistical model. For instance
it works only with a linear model.

"year" corresponds to an average year of 365.25 days and month to an average month of 30,4575 days.
}
\author{
P. Chevallier - January 2024
}
