% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_change_tz.R
\name{h_changetz}
\alias{h_changetz}
\title{Change the time zone of a time series}
\usage{
h_changetz(file, tz1 = "UTC", tz2 = "Europe/Paris")
}
\arguments{
\item{file}{File name to proceed}

\item{tz1}{original time zone (default = "UTC")}

\item{tz2}{new time zone (default = "Europe/Paris")}
}
\description{
Change the time zone of a time series
}
\details{
The output file is named with a tz prefix.
}
\author{
P. Chevallier - June 2023
}
