% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_common.R
\name{h_common}
\alias{h_common}
\title{Extract 2 (or more) time-series on their common period}
\usage{
h_common(files)
}
\arguments{
\item{files}{List of file names to process.}
}
\value{
hts files resulting of the operation; their names are composed as:
co_<original filename>
}
\description{
The fonction extract the data of 2 (or more) hts time-series
for the common date/time records (precision of the second).
}
\examples{
\dontrun{
f <- h_common(files = c("foo1.hts","foo2.hts"))
}
}
\author{
P. Chevallier - Oct 2017-June 2023
}
