% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_condition.R
\name{h_condition}
\alias{h_condition}
\title{Conditional extraction of a time-series regarding another one}
\usage{
h_condition(files, condition)
}
\arguments{
\item{files}{Liste de 2 file names}

\item{condition}{Liste 3 objects : oper ("sup" or "inf" or "between"),
thrhd1 < thrhd2 ; default is c("inf",0,NA)}
}
\description{
The series to proceed is the first of the list, the conditional series
the second. Only the common record dates are kept.
}
\details{
If the condition on the file 2 value is not respected, the value of
file 1 is changed as NA.

The condition has 3 options : x< ("inf"), x>= ("sup"), < x <=
("between"). In case of errror or by default, "inf" is considered.
In the cases "inf" and "sup", only one threshold is used (thrhd1) ;
in the case "between", two theresholds are needed (thrhd1 < thrhd2).

The output file is the name of the fist file with a cd_ prefix.
}
\examples{
\dontrun{

f <- h_condition(c(f1,f2), c("between", O, 2))
}

}
\author{
P. Chevallier - Oct 2017-Jan 2019
}
