% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_cumul.R
\name{h_cumul}
\alias{h_cumul}
\title{Cumul of time-series}
\usage{
h_cumul(file, start = NA, end = NA)
}
\arguments{
\item{file}{File name to proceed}

\item{start}{Start date, default = NA}

\item{end}{End date, default NA}
}
\description{
The function returns a time-series of cumulated values. If the value
is negative, the absolute value is taken.
It is possible to limit the computation time interval. NA values are ignored.
}
\details{
The output file is named with a cu_ prefix.
}
\examples{
\dontrun{

f <- h_cumul(f, start="2012-1-1", end = "2013-1-1")
}

}
\author{
P. Chevallier - Oct 2017-Jan 2019
}
