% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_gaperr.R
\name{h_gaperr}
\alias{h_gaperr}
\title{Replace errors with gaps in a time-series based on neighboring values}
\usage{
h_gaperr(file, nv = 1, itv0 = 43201, df)
}
\arguments{
\item{file, }{File name to proceed}

\item{nv}{Number of below and above neighboring values to take into account,
default = 1}

\item{itv0}{Threshold of minimum time gap (see function h_gaprem_itv)}

\item{df}{Deviation value factor for testing if a value is correct or not}
}
\value{
a time-series file with the prefix eg_
}
\description{
Replace errors with gaps in a time-series based on neighboring values
}
\details{
Replace errors with gaps in a time-series based on neighboring values
}
\author{
P. Chevallier - Nov 2019
}
