% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_gapfill.R
\name{h_gapfill}
\alias{h_gapfill}
\title{Simple gapfilling in a time-series}
\usage{
h_gapfill(file, npdt)
}
\arguments{
\item{file}{File name to proceed}

\item{npdt}{Number of time-steps}
}
\value{
a time-series file with the prefix gf_
}
\description{
Simple gapfilling in a time-series
}
\details{
Replace the missing values with the linear interpolated value within the gap
interval, when the time interval is less than a number of fixed time steps.

CAUTION! this operation is only possible when the time-series has a fixed time-step.
}
\author{
P. Chevallier - Nov 2017 - Nov 2021
}
