% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_gaprem_itv.R
\name{h_gaprem_itv}
\alias{h_gaprem_itv}
\title{Remove gaps in a time-series with a time interval threshold}
\usage{
h_gaprem_itv(file, itv0 = 43201)
}
\arguments{
\item{file}{File name to proceed}

\item{itv0}{Time threshold in seconds, default = 43201 (i.e 12 hours)}
}
\value{
a time-series file with the prefix gr_
}
\description{
Remove gaps in a time-series with a time interval threshold
}
\details{
Remove the missing values when the time interval between the previous
and the next record is less than a fixed threshold
}
\author{
P. Chevallier - Nov 2019
}
