% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_rainsnow.R
\name{h_rainsnow}
\alias{h_rainsnow}
\title{Share the solid and liquid precipitations with a temperature criteria}
\usage{
h_rainsnow(fpr, fta, ta0, ta1, sta = NA)
}
\arguments{
\item{fpr}{Precipitation file name}

\item{fta}{Temperature file name}

\item{ta0}{Low temperature threshold}

\item{ta1}{High temperature threshold}

\item{sta}{Station id. (default = NA)}
}
\value{
2 hts files, one with the liquid precipitation (prefix rn_) and one with
the solid precipitation (prefix sn_).
}
\description{
The precipitations are shared with a linear bevel between two temperature
 values
}
\details{
The two time-series must be previously restricted to the same interval of time.

The two temperature thresholds can be equal.

The temperature time-series must be complete with no gap. Gaps are allowed
in the precipitation time-series.

Is the station id is NA, the station id of the file fta is used.
}
\author{
P. Chevallier - Oct 2017- Feb 2019
}
