% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_replace.R
\name{h_replace}
\alias{h_replace}
\title{Replace a value by another}
\usage{
h_replace(file, old.val, new.val)
}
\arguments{
\item{file}{File name to proceed}

\item{old.val}{Value to be replaced}

\item{new.val}{New value}
}
\description{
Replace a value by another
}
\details{
The output file is named with a re_ prefix.
}
\examples{
\dontrun{

f <- ts_remplace_ts(f,  NA, 0)
}

}
\author{
P. Chevallier - Oct 2017- Nov 2020
}
