% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_restrict.R
\name{h_restrict}
\alias{h_restrict}
\title{Restrict a series between 2 dates}
\usage{
h_restrict(file, start = NA, end = NA)
}
\arguments{
\item{file}{File name to proceed}

\item{start}{Start date/time (included) of POSIXct class (default = NA)}

\item{end}{End date/time (excluded) of POSIXct class (default = NA)}
}
\description{
Restrict a series between 2 dates
}
\details{
The output file is named with a rs_ prefix.
}
\author{
P. Chevallier - Nov 2017-Jan 2019
}
