% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_rollav.R
\name{h_rollav}
\alias{h_rollav}
\title{Rolling average of a daily time-series}
\usage{
h_rollav(file, ti = 7, position = "central")
}
\arguments{
\item{file}{File name to proceed}

\item{ti}{Time interval of computation in days (default = 7)}

\item{position}{Position "central" or "right"}
}
\description{
The function compute a rollong average of daily time-series values.
NA values are removed.
}
\details{
The output file is named with a ro_ prefix. The computation can considers
the values before and after the current time step (position = "central") or the values
before the current time step. If the position is "central", the position must be an
odd integer.
}
\author{
P. Chevallier - Apr 2020
}
