% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_stat_basic.R
\name{h_stat_basic}
\alias{h_stat_basic}
\title{Basic statistics of a time-series}
\usage{
h_stat_basic(files)
}
\arguments{
\item{files}{vector of file names to process}
}
\value{
a tibble with the basic stats of the files.
}
\description{
Compute the main statistic parameters of a time-series
}
\examples{
\dontrun{
simplestat <- h_stat_basic(c("foo1.hts", "foo2.hts")
}

}
\author{
P. Chevallier - Oct 2017 - Feb 2022
}
