% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_substitute.R
\name{h_substitute}
\alias{h_substitute}
\title{Subtitute the missing values in a series by existing values of another series}
\usage{
h_substitute(files)
}
\arguments{
\item{files}{List of two file names}
}
\description{
The series to proceed (first in file list) contents missing values or gaps
to be replaced by those of the second series (second in file list).

The function only works on the common dates of both series.
}
\details{
The output file is named with a sb_ prefix.
}
\examples{
\dontrun{

f <- h_substitute(c(f1, f2))
}

}
\author{
P. Chevallier - Feb 2017 - Mar 2020
}
