% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htsr-package.R
\docType{package}
\name{htsr-package}
\alias{htsr}
\alias{htsr-package}
\title{htsr: Hydro-Meteorology Time-Series}
\description{
Functions for the management and treatment of hydrology and meteorology time-series stored in a 'Sqlite' data base.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/p-chevallier/htsr}
  \item Report bugs at \url{https://github.com/p-chevallier/htsr/issues}
}

}
\author{
\strong{Maintainer}: Pierre Chevallier \email{pierre.chevallier@mailo.com}

Authors:
\itemize{
  \item Pierre Chevallier \email{pierre.chevallier@mailo.com}
}

}
\keyword{internal}
