% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htsr.R
\docType{package}
\name{htsr}
\alias{htsr}
\title{htsr package}
\description{
htsr: A package for managing sqlite data bases, which contain hydro-meteorology
time-series.

The htsr package contains 6 types of functions: data base (d_<NAME>), file
f_<NAME>, hydromet (h_<NAME>), plotting (p_<NAME>), shiny (s_<NAME>) and
miscellaneous (z_<NAME).
}
