% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_box_month.R
\name{p_box_month}
\alias{p_box_month}
\title{Boxplot of the 12 months of a time-series.}
\usage{
p_box_month(
  file,
  title = "Title",
  axeY = "Y-axis",
  savefig = FALSE,
  fileo = "plot.png",
  width = 8,
  height = 6
)
}
\arguments{
\item{file}{File name of the time-series}

\item{title}{Title plot (default = Title)}

\item{axeY}{Title of y-axis (default  Y-axis)}

\item{savefig}{Save plot file TRUE / FALSE (default)}

\item{fileo}{Name of the plot file with extension png, jpg or pdf}

\item{width}{Plot width (x 100 pixels), default = 8}

\item{height}{Plot heights (x 100 pixels), default = 6}
}
\value{
A ggplot2 object
}
\description{
Boxplot of the 12 months of a time-series.
}
\author{
P. Chevallier - Nov 2017 -Feb 2019
}
