% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_plothts.R
\name{p_plothts}
\alias{p_plothts}
\title{Shiny app: plot hts files}
\usage{
p_plothts()
}
\description{
Shiny application of the \code{\link{p_line_app}} and \code{\link{p_bar_app}}
functions
}
\details{
When launched, this function first ask to give the number of files to plot
and to select them. After that a shiny window is open.
 The setting tab must be first completed and saved. Then go to the plot tab, complete
 the plotting settings and press Plot.
 If you want to modify the initial settings, return to the setting tab.
 After setting changes
 press Save one more time, go to the plot tab and press Plot!

When finished, press Done to exit from the shiny windows
}
\author{
P. Chevallier - May 2020
}
