% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{u_index}
\alias{u_index}
\title{Compute an index of community}
\arguments{
\item{nz}{length of the concatenated time-series}

\item{yd}{initial vector of datetimes (in sec)}
}
\value{
vector of indexes
}
\description{
Compute an index of community
}
\details{
the function compute an index, which the number of apparition of the same datetime
in a time-series
}
\author{
P. Chevallier - April 2023
}
