% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{u_timestep}
\alias{u_timestep}
\title{Compute values in a time-series with a fixed timestep}
\arguments{
\item{te}{time end (in sec)}

\item{yd}{initial vector of datetimes (in sec)}

\item{yv}{initial vector of values}

\item{tst}{timestep (in mn)}

\item{iop}{operation index}
}
\value{
vector of values with fixed timestep
}
\description{
Compute values in a time-series with a fixed timestep
}
\details{
iop = 1 for sum; 0 for mean; -2 for min and +2 for max
}
\author{
P. Chevallier - June 2023
}
