% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_temp_alt.R
\name{w_temp_alt}
\alias{w_temp_alt}
\title{Compute temperature, function of altitude}
\usage{
w_temp_alt(file, alt0 = 0, alt, grad = -0.0065)
}
\arguments{
\item{file}{File name of the known air temperature ts (°C)}

\item{alt0}{Altitude of the known air temperature ts - default = 0 (m)}

\item{alt}{Altitude of the computed air- temperature ts (m)}

\item{grad}{Temperature gradient vs elevation - default = -0.0065 (°C/m)}
}
\value{
An hts file with the suffix _<alt>
}
\description{
Compute temperature, function of altitude
}
\details{
The function computes an air temperature time-series at a given altitude,
based on a known air temperature time-series at a known altitude.
}
\author{
P. Chevallier - Nov 2021
}
