% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_coord.R
\name{z_coord}
\alias{z_coord}
\title{Coordinate utility}
\usage{
z_coord(ncoord = NA, ccoord = NA, type)
}
\arguments{
\item{ncoord}{Numeric coordinate}

\item{ccoord}{Character coordinate}

\item{type}{Lat / Lon}
}
\value{
Coordinates in characters
}
\description{
Convert numeric coordinates in character coordinates
}
\details{
Only one of both parameters ncoord (numeric) and ccoord (character) must be filled,
the other one remaining NA. The type of coordinate (Lat or Lon) is compulsory.

The character coordinate must be organized in one string with 4 fields (degrees,
minutes, seconds, direction) separated with blanks (space or tab). Within each field,
no blanks are allowed to share the numeric value and the unit character. For
the unit character,  the only following letters are allowed: letter d/m/s.
For direction, the only the following letters are allowed: N/n/W/w/S/s/E/e.

Example: "25d 18m 56.2s S"
}
\author{
P. Chevallier - Jan 2019 / Nov 2020
}
