% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_convert_meteofrance_d.R
\name{d_convert_meteofrance_d}
\alias{d_convert_meteofrance_d}
\title{Convert a Meteo-France csv daily basic data file into a htsr sqlite base}
\usage{
d_convert_meteofrance_d(fmeteo)
}
\arguments{
\item{fmeteo}{Full name of the Meteo-France csv file}
}
\description{
Convert a Meteo-France csv daily data file into a htsr sqlite base. It regards
the "basic" data file, which includes precipitation, temperature and wind data. For other
variables the function d_convert_meteofrance_d1 shall be used with the corresponding csv file.
The csv file shall be downloaded from https://meteo.data.gouv.fr/
The name of the created sqlite file is the same as the csv file with an extension .sqlite.
}
\details{
The sensors have an additional prefix d (as daily) in order to distinguish them from sensors with another
time reference.
}
\author{
P. Chevallier - dec 2023 - jan 2024
}
