% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_convert_meteofrance_h.R
\name{d_convert_meteofrance_h}
\alias{d_convert_meteofrance_h}
\title{Convert a Meteo-France csv hourly basic data file into a htsr sqlite base}
\usage{
d_convert_meteofrance_h(fmeteo)
}
\arguments{
\item{fmeteo}{Full name of the Meteo-France csv file}
}
\description{
Convert a Meteo-France csv hourly data file into a htsr sqlite base. It regards
the "basic" data file, which includes precipitation, temperature and wind data.
The csv file shall be downloaded from https://meteo.data.gouv.fr/
The name of the created sqlite file is the same as the csv file with an extension .sqlite.
}
\details{
The sensors have an additional prefix h (as hourly) in order to distinguish them from sensors with another
time reference.
}
\author{
P. Chevallier - jan 2024
}
