% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_convert_weewx.R
\name{d_convert_weewx}
\alias{d_convert_weewx}
\title{Convert a weewx data base into a htsr sqlite base}
\usage{
d_convert_weewx(db.weewx, fsq = NA, update = TRUE, sta = NA, name_st = NA)
}
\arguments{
\item{db.weewx}{Full name of the weewx data base}

\item{fsq}{Full name of the htsr data base}

\item{update}{(default = TRUE)}

\item{sta}{Station id (default = NA)}

\item{name_st}{Station name (default = NA)}
}
\description{
Convert (or update) a weewx data base into a htsr sqlite base
}
\details{
If update is TRUE, sta and name_st are unnecessary. I update is FALSE and
fsq is NA, fsq is named "weewx.sqlite".
}
\examples{
\dontrun{

d_convert_weewx("weewx.sql", "foo.sqlite")
}

}
\author{
P. Chevallier - Feb 2018 - Sep 2023
}
