% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_convert_weewx.R
\name{d_convert_weewx}
\alias{d_convert_weewx}
\title{Convert a weewx data base into a htsr sqlite base}
\usage{
d_convert_weewx(db.weewx, fsq, sta, name_st)
}
\arguments{
\item{db.weewx}{Full name of the weewx data base}

\item{fsq}{Full name of the htsr data base}

\item{sta}{Station id}

\item{name_st}{Station name}
}
\description{
Convert a weewx data base into a htsr sqlite base
}
\examples{
\dontrun{

d_convert_weewx("weewx.sql", "foo.sqlite")
}

}
\author{
P. Chevallier - Feb 2018 - Jun 2023
}
