% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_exp_discalib.R
\name{d_exp_discalib}
\alias{d_exp_discalib}
\title{Export discharge measurements and calibrations from data base}
\usage{
d_exp_discalib(fsq, sta, calib = TRUE, dism = TRUE)
}
\arguments{
\item{fsq}{Full name of the data base}

\item{sta}{Station Id.}

\item{calib}{Calibration extraction TRUE (default)/FALSE}

\item{dism}{Discharge measurement extraction TRUE (default)/FALSE}
}
\value{
a list of 2 tibbles, one with the calibration table and one with the discharge measurements
}
\description{
Export discharge measurements and calibrations from data base
}
\seealso{
\code{\link{ds_exp_hts}} for export time-series
}
\author{
P. Chevallier - Sep 2017 - Nov 2020
}
