% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_inventory.R
\name{ds_inventory}
\alias{ds_inventory}
\title{Shiny app: inventory of htsr sqlite data base}
\usage{
ds_inventory()
}
\value{
a shiny session
}
\description{
Shiny application of the \code{\link{d_inventory}} function
}
\details{
Complete the requested information in the left panel, then press the submit button.
 If the station field is empty, the function will return the list of the stations
 in the data base. If the station field is filled, the function will return the
 list of the station sensors in the data base. When finished press "done".
}
\author{
P. Chevallier - Sep 2020 - Sep 2023
}
