% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_inventory.R
\name{ds_inventory}
\alias{ds_inventory}
\title{Shiny app: inventory of htsr sqlite data base}
\usage{
ds_inventory(fsq)
}
\arguments{
\item{fsq}{File name of the Sqlite data base}
}
\value{
a shiny session
}
\description{
Shiny application of the \code{\link{d_inventory}} function
}
\details{
Complete the requested information in the left panel, then press the submit button.
 If the station field is empty, the function will return the list of the stations
 in the data base. If the station field is filled, the function will return the
 list of the station sensors in the data base. When finished press "done".

If the data do not exist, the app crashes and error messages
are displayed in the console window.
}
\author{
P. Chevallier - Sep-Nov 2020
}
