% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_sensor.R
\name{ds_sensor}
\alias{ds_sensor}
\title{Shiny app: create, modify or remove a sensor from a data base}
\usage{
ds_sensor()
}
\value{
a shiny session
}
\description{
Shiny application of the \code{\link{d_sensor}} function
}
\author{
P. Chevallier - Nov 2020 - Sep 2023
}
