% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_station.R
\name{ds_station}
\alias{ds_station}
\title{Shiny app: create, modify or remove a station from a data base}
\usage{
ds_station(fsq)
}
\arguments{
\item{fsq}{File name of the Sqlite data base}
}
\value{
a shiny session
}
\description{
Shiny application of the \code{\link{d_station}} function
}
\author{
P. Chevallier - Apr-Nov 2020
}
