% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_gapfill.R
\name{h_gapfill}
\alias{h_gapfill}
\title{Simple gapfilling in a time-series}
\usage{
h_gapfill(file, npdt)
}
\arguments{
\item{file}{File name to proceed}

\item{npdt}{Number of time-steps}
}
\value{
a time-series file with the prefix gf_
}
\description{
Simple gapfilling in a time-series
}
\details{
Repalce the missing values with the mean of the last and next values
when the time interval is less than a number of fixed time steps.
}
\author{
P. Chevallier - Nov 2017 - Feb 2019
}
