% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_stat_basic.R
\name{h_stat_basic}
\alias{h_stat_basic}
\title{Basic statistics of a time-series}
\usage{
h_stat_basic(file)
}
\arguments{
\item{file}{File to process}
}
\value{
nb_val, mean, standard deviation, min, quantile .25, mediane, quantile .75, max
and display a boxplot with these values.
}
\description{
Compute the main statistic parameters of a time-series
}
\examples{
\dontrun{

simplestat <- h_stat_basic("foo.hts")
}

}
\author{
P. Chevallier - Oct 2017-Jan 2019
}
