% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_tstep.R
\name{hs_tstep}
\alias{hs_tstep}
\title{Shiny app: convert f file with fixed time-step}
\usage{
hs_tstep(file)
}
\arguments{
\item{file}{File name of the Sqlite data base}
}
\value{
a shiny session
}
\description{
Shiny application of the  functions \code{\link{h_timestep}} and \code{\link{h_month}}
}
\author{
P. Chevallier - Dec 2020 - Oct 2021
}
