% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_bar.R
\name{p_bar}
\alias{p_bar}
\title{Bar plot}
\usage{
p_bar()
}
\value{
a ggplot2 object
}
\description{
Bar plot based on htsr time-series. Prior to the execution of the function,
the parameters must be set by a function generating a "settings.RData" file in the working 
directory: \code{\link{z_set}} or \code{\link{ps_plothts}}.
}
\details{
The function deosn't have parameter. For a full description of the settings, see \code{\link{z_set}}
}
\seealso{
\code{\link{p_line}} for plotting bars and
 \code{\link{z_set}} for setting the plot parameters
}
\author{
P. Chevallier - Apr 2015 - Aug 2023
}
