% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_plothts.R
\name{ps_plothts}
\alias{ps_plothts}
\title{Shiny app: plot hts files}
\usage{
ps_plothts(files)
}
\arguments{
\item{files}{List of the time-series files to be plotted.}
}
\description{
Shiny application of the \code{\link{p_line_app}} and \code{\link{p_bar_app}}
functions
}
\details{
When launched, a shiny window is open.
 The setting tab must be first completed and saved. Then go to the plot tab, complete
 the plotting settings and press Plot.
 If you want to modify the initial settings, return to the setting tab.
 After setting changes
 press Save one more time, go to the plot tab and press Plot!

When finished, press Done to exit from the shiny windows
}
\author{
P. Chevallier - May 2020
}
