% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_plothts.R
\name{ps_plothts}
\alias{ps_plothts}
\title{Shiny app: plot hts files}
\usage{
ps_plothts()
}
\description{
Shiny application of the \code{\link{p_line}} and \code{\link{p_bar}}
functions, associated with \code{\link{z_set}}
}
\details{
When launched, a shiny window is open. Follow the instructions and steps.
}
\author{
P. Chevallier - May 2020 - Aug 2023
}
