% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_set.R
\name{z_set}
\alias{z_set}
\title{Edit settings}
\usage{
z_set(tz = FALSE, mapal = FALSE)
}
\arguments{
\item{tz}{Logical, setting time zone (default=FALSE)}

\item{mapal}{Logical, setting my palette (default=FALSE)}
}
\description{
Utility for editing the settings of the htsr package.
}
\details{
The function allows to edit user settings for time zone and color palette.

The settings are stored in an external data file of the htsr package,
named "settings.RData".

tz is the time zone coded following the Olson standard list.

my palette is a list of 12 colors from the R color name list.
}
\author{
P. Chevallier - nov 2018 - nov 2020
}
